alzheimer-mri-assistant/requirements.txt
fastapi>=0.115,<0.116
uvicorn>=0.30,<0.31
numpy>=1.26,<2.3
opencv-python>=4.10,<4.11
pillow>=10.4,<11
onnxruntime>=1.19,<1.20
tensorflow>=2.17,<2.19
python-multipart>=0.0.9,<0.0.10
pydantic>=2.9,<2.10
